Examples/FunctionApps/GlupXorFunctionApp/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
tensorflow==2.5.0
numpy==1.19.2
